# api/openapi.yaml
openapi: 3.1.0
info:
  title: Parsley API
  description: ユーザー情報の管理を行うAPI（IDはDB側で管理）。
  version: 1.0.0

servers:
  - url: http://localhost:8080

paths:
  /users:
    get:
      summary: 全ユーザーの取得
      operationId: listUsers
      responses:
        '200':
          description: ユーザー一覧
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      summary: 新規ユーザーの登録
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: 登録完了
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  /users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: uint # GORMのuint型に合わせる
    get:
      summary: 特定ユーザーの取得
      operationId: getUserById
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    put:
      summary: 登録内容の変更
      operationId: updateUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    delete:
      summary: ユーザーの削除
      operationId: deleteUser
      responses:
        '204':
          description: 削除成功

components:
  schemas:
    User:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
          example: "パセリ太郎"
        email:
          type: string
          format: email
          example: "parsley@example.com"